yolov8/yolov8_python/requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.6.0
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
huggingface_hub

# Logging -------------------------------------
tensorboard>=2.4.1
clearml
comet

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
onnxruntime
coremltools>=6.0  # CoreML export
onnx>=1.12.0  # ONNX export
onnx-simplifier>=0.4.1  # ONNX simplifier
scikit-learn==0.19.2  # CoreML quantization
tensorflow>=2.4.1  # TF exports (-cpu, -aarch64, -macos)
tensorflowjs>=3.9.0  # TF.js export

# Extras --------------------------------------
ipython  # interactive notebook
ipykernel #python kernel for notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation
albumentations>=1.0.3
pycocotools>=2.0.6  # COCO mAP
roboflow
